Refactor BE to support Long ids instead of int. Fixed file upload and post creation

+++ BE/src/main/java/com/g3/feedbackApp/DataSources/FakeDataSourcePost.java
@@ -5,37 +5,35 @@ import com.g3.feedbackApp.DataSources.Interfaces.IDataSourceReviewer;
 import com.g3.feedbackApp.Models.ReviewerModel;
 import com.g3.feedbackApp.Models.PostModel;
 import com.g3.feedbackApp.Models.VersionModel;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Repository;
 
 import java.nio.file.Path;
 import java.time.LocalDate;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
 @Repository
 public class FakeDataSourcePost implements IDataSourcePost {
 
-    private static int postIdCounter = 4;
+    private static Long postIdCounter = 4L;
     List<PostModel> postModelList = new ArrayList<>();
 
-    private static int versionIdCounter = 4;
+    private static Long versionIdCounter = 4L;
     List<VersionModel> versionModelList = new ArrayList<>();
 
-    private static int reviewerConnIdCounter = 1;
+    private static Long reviewerConnIdCounter = 1L;
     List<ReviewerModel> reviewerModelList = new ArrayList<>();
 
     private IDataSourceReviewer dataSourceReviewer = new FakeDataSourceReviewer();
 
     public FakeDataSourcePost(){
-        postModelList.add(new PostModel(1, 1,"First Post", "code", "my first code post", LocalDate.now(), null));
-        postModelList.add(new PostModel(2, 1,"Second Post", "Assay", "Please check my assay", LocalDate.now(), null));
-        postModelList.add(new PostModel(3, 1,"Third Post", "DB Diagrams", "The first draft", LocalDate.now(), null));
+        postModelList.add(new PostModel(1L, 1,"First Post", "code", "my first code post", LocalDate.now(), null));
+        postModelList.add(new PostModel(2L, 1,"Second Post", "Assay", "Please check my assay", LocalDate.now(), null));
+        postModelList.add(new PostModel(3L, 1,"Third Post", "DB Diagrams", "The first draft", LocalDate.now(), null));
 
-        versionModelList.add(new VersionModel(1, 1, Path.of("testFilePathVersion1")));
-        versionModelList.add(new VersionModel(2, 1, Path.of("testFilePathVersion2")));
-        versionModelList.add(new VersionModel(3, 1, Path.of("testFilePathVersion3")));
+        versionModelList.add(new VersionModel(1L, 1L, Path.of("testFilePathVersion1")));
+        versionModelList.add(new VersionModel(2L, 1L, Path.of("testFilePathVersion2")));
+        versionModelList.add(new VersionModel(3L, 1L, Path.of("testFilePathVersion3")));
 
 //        reviewerModelList.add(new ReviewerModel(1, 1, 2));
 //        reviewerModelList.add(new ReviewerModel(2, 2, 2));
@@ -54,23 +52,23 @@ public class FakeDataSourcePost implements IDataSourcePost {
     }
 
     @Override
-    public boolean createVersion(int postId, Path filePath) {
+    public boolean createVersion(Long postId, Path filePath) {
         versionModelList.add(new VersionModel(versionIdCounter, postId, filePath));
         versionIdCounter++;
         return true;
     }
 
     @Override
-    public boolean assignReviewers(List<Integer> reviewersIds, int postId) {
-        for(int id: reviewersIds){
-            reviewerModelList.add(new ReviewerModel(reviewerConnIdCounter, postId, id));
+    public boolean assignReviewers(List<Long> reviewersIds, Long postId) {
+        for(Long id: reviewersIds){
+            reviewerModelList.add(new ReviewerModel(reviewerConnIdCounter, postId, id.intValue()));
             reviewerConnIdCounter++;
         }
         return true;
     }
 
     @Override
-    public PostModel getPostWithId(int Id) {
+    public PostModel getPostWithId(Long Id) {
         for(PostModel p: postModelList){
             if(p.getPostId() == Id){
                 return p;
@@ -80,7 +78,7 @@ public class FakeDataSourcePost implements IDataSourcePost {
     }
 
     @Override
-    public VersionModel getVersionWithId(int versionId) {
+    public VersionModel getVersionWithId(Long versionId) {
         for(VersionModel v: versionModelList){
             if(v.getVersionId() == versionId){
                 return v;
@@ -90,7 +88,7 @@ public class FakeDataSourcePost implements IDataSourcePost {
     }
 
     @Override
-    public List<VersionModel> getVersionsForPost(int postId) {
+    public List<VersionModel> getVersionsForPost(Long postId) {
         List<VersionModel> listOfIds = new ArrayList<>();
         for(VersionModel v: versionModelList){
             if(v.getPostId() == postId){
@@ -101,8 +99,8 @@ public class FakeDataSourcePost implements IDataSourcePost {
     }
 
     @Override
-    public List<Integer> getReviewersIdsForPost(int postId) {
-        List<Integer> listOfIds = new ArrayList<>();
+    public List<Long> getReviewersIdsForPost(int postId) {
+        List<Long> listOfIds = new ArrayList<>();
         for(ReviewerModel reviewer: reviewerModelList){
             if(reviewer.getPostId() == postId){
                 listOfIds.add(reviewer.getId());
